// src/modules/auth/auth.dto.ts
import { z } from "zod";

export const signupSchema = z.object({
    email: z.string().email(),
    password: z.string().min(8),
    name: z.string().min(1),
});

export type SignupDto = z.infer<typeof signupSchema>;

export const loginSchema = z.object({
    email: z.string().email(),
    password: z.string().min(1),
});

export type LoginDto = z.infer<typeof loginSchema>;